refactor: IInput 컴포넌트의 Storybook 정의 개선 및 확장
